Planora/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
decorator==5.2.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0
frozenlist==1.7.0
genai==2.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1rc0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.20.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.1.7
multidict==6.4.4
numpy==2.2.6
openai==0.27.10
packaging==25.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tiktoken==0.3.3
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
Twisted==25.5.0
txaio==23.1.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
zope.interface==7.2
